# [Mac] Set Marshal*Exception modes