This should prevent hard crashing when working with exception handling in xammac

Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1042626