@Therzok 2019-12-03
Testing perf impact myself, to see if it's actually negligible