@Therzok 2019-12-03
Attached are the 2 spindump files, the version with this patch is the one at the 01:21 mark
[Visual Studio_Profile_2019-12-04__01-21-33.txt](https://github.com/mono/monodevelop/files/3919223/Visual.Studio_Profile_2019-12-04__01-21-33.txt)
[Visual Studio_Profile_2019-12-04__01-15-25.txt](https://github.com/mono/monodevelop/files/3919224/Visual.Studio_Profile_2019-12-04__01-15-25.txt)